requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pymongo
python-dotenv
